getAll implemented and delete duplicate files.
