+++ src/main/java/todolist_grupo3/controllers/TaskController.java
@@ -1,0 +1,28 @@
+package todolist_grupo3.controllers;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.AllArgsConstructor;
+import todolist_grupo3.services.TaskService;
+
+
+@RestController
+@RequestMapping("/todolist")
+@AllArgsConstructor
+public class TaskController {
+    @Autowired
+    private TaskService taskService;
+
+    @GetMapping("/getAll")
+    @CrossOrigin("*")
+    public ResponseEntity<?> getAll() {
+        return ResponseEntity.status(HttpStatus.OK).body(taskService.getAllTasks());
+    }
+
+}
