+++ src/main/java/todolist_grupo3/entities/State.java
@@ -1,0 +1,5 @@
+package todolist_grupo3.entities;
+
+public enum State {
+    COMPLETE, INCOMPLETE;
+}
