+++ src/main/java/todolist_grupo3/services/TaskService.java
@@ -1,0 +1,9 @@
+package todolist_grupo3.services;
+
+import java.util.List;
+
+import todolist_grupo3.entities.Task;
+
+public interface TaskService {
+        List<Task> getAllTasks();
+}
