+++ src/main/java/todolist_grupo3/services/TaskServiceImpl.java
@@ -1,15 +1,23 @@
-package todolist_grupo3.todolist_grupo3.services;
+package todolist_grupo3.services;
+
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import lombok.AllArgsConstructor;
-import todolist_grupo3.todolist_grupo3.repo.TaskRepository;
+import todolist_grupo3.entities.Task;
+import todolist_grupo3.repo.TaskRepository;
 
 @Service
 @AllArgsConstructor
 public class TaskServiceImpl implements TaskService{
     @Autowired
     private TaskRepository taskRepository;
+
+    @Override
+    public List<Task> getAllTasks(){
+        return taskRepository.findAll();
+    }
 
 }
