--- src/main/java/todolist_grupo3/todolist_grupo3/controllers/TaskController.java
@@ -1,17 +1,0 @@
-package todolist_grupo3.todolist_grupo3.controllers;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.AllArgsConstructor;
-import todolist_grupo3.todolist_grupo3.services.TaskService;
-
-
-@RestController
-@RequestMapping("/todolist")
-@AllArgsConstructor
-public class TaskController {
-    @Autowired
-    private TaskService taskService;
-}
