--- src/main/java/todolist_grupo3/todolist_grupo3/entities/State.java
@@ -1,5 +1,0 @@
-package todolist_grupo3.todolist_grupo3.entities;
-
-public enum State {
-    COMPLETE, INCOMPLETE;
-}
