--- src/main/java/todolist_grupo3/todolist_grupo3/services/TaskService.java
@@ -1,5 +1,0 @@
-package todolist_grupo3.todolist_grupo3.services;
-
-public interface TaskService {
-
-}
